Some fixes and license

+++ src/compiler_stubs.c
@@ -1,21 +1,21 @@
 #include "stdtypes.h"
 
-void* memset(void* dest, int c, u32 n);
+void* memset(void* dest, int c, u64 n);
 void* memset
-(void* dest, int c, u32 n)
+(void* dest, int c, u64 n)
 {
-    u32 i = 0;
+    u64 i = 0;
     for(;i < n; ++i) {
         ((u8*)dest)[i] = (u8)c;
     }
     return dest;
 }
 
-void* memcpy(void* dest, const void* src, uint32_t n);
+void* memcpy(void* dest, const void* src, u64 n);
 void* memcpy
-(void* dest, const void* src, uint32_t n)
+(void* dest, const void* src, u64 n)
 {
-    u32 i = 0;
+    u64 i = 0;
     for(;i < n; ++i) {
         ((u8*)dest)[i] = ((u8*)src)[i];
     }
